Add GitHub action setup for ubuntu
This commit adds the GitHub actions setup to build and run our
suite on ubuntu for now. Currently, it only usages one single
ruby version to run our test suite.

As soon as the gem is ready then let's also adds other platforms
and also add multiple ruby version so we can be sure it's stable
in our expected eco-system.