Add support to copy font to a path
We want fontist to take care of copying and downloading a fonts
when necessary. This commit takes the first step to copy fonts
to the user specified paths.

Ideally, fontist will take care of fonts lookup or download a
font. But this commit primarily focuses on looking into some of
the default font's locations with added support for a custom
directory. We currently organized these directories data based
on the core operating system, but we will extend those as soon
as it become necessary.