Add an interface to find or download a font
This commit adds an interface that will allow us to find a font in
the user's system. Or even better, if the font is not within user's
system but one of our supported font then it will download the font
and return the fonts path to the user. Usages:

```ruby
Fontist::Finder.find("CALIBRI.TTF")
```