Adjust font lookup and downloading
Currently, the font lookup is case sensitive, and it also does not use
any of the caches. So, this commit changes this and it will try use the
case insensitive version for font lookup.

This also usages the cache unless we explicitly specify to download the
file again. And, the font lookup also includes the `assets` directory,
this will also be used when finding a font.