Do not download the font through finder
Currently, the finder interface is doing the job for font lookup
and also to download the fonts if missing. But, in reality we need
user to accept the font terms, so this commit changes this and it
will raise an error for installable fonts instead of downloading
it directly.

Later, we will also add suggestion for user on how to download
those font from the supported libraries.