Add installer to download support font
For uninstalled font we want to allow user to download those with
proper permission. This commit adds those necessary changes, so now
user have an way to utilize fontist to download any of the supported
fonts. This interface will also return the downloaded paths, so the
user can decide what to do with that.