Add action to run test suite in windows
This commit adds the github action setup to run the test suite
in windows. This commit also adds the correct path for windows
font lookup.

Since, all of these fonts are already available in widows so
we also skipped the download process for that. But if there
are some edge cases then we will also need to take care of
those in the future.