Add support for meta data for fonts
As of now, we are doing hard lookup for fonts by their exact
name and also there are no way to specify font meta data like
human readable name or different styles.

But in the long term plan - moving forward with font formulas
we need to specify font style as well. So, this commit adds
those necessary changes so we can specify font name, and the
styles for font. This also adopt this changes for the lookup