Adopt font lookup for font meta data
Previously, we were only using the direct font name for lookup
and installation, so it was pretty straight forward. But, now
that font lookup/installation using font meta data like, name
so this commit adopt those changes.

This changes the system font interface and as well as the base
formula interface, so the library should work as expected.