Remove formulas from the old structure
Since we are going to import all formulas from the fontist
formulas repository, so having the old formulas might create
some confliction.

So, this commit remove those for now, and keep it as a new
pull request so we can easily reuse some of the stuff and
a place to come back in something goes south.