Simplify the font finding interface
Currently we've a `Fontist::Finder` interface to find a font,
which is little bit verbose as it doesn't exactly tell what is
it gonna find, a font or a formula?

This commit changes this interface to `Fontist::Font`, so now
user can be sure what they are dealing with. This also adds the
benefits of having a `install` or `list` interface in future.