Remove unused temporary asset path
Since we've created the `~/.fontist` global directory we are not
using the `assets/fonts` folder anymore. But there are some tests
that's still using the folder for temporary assets.

This commit removes this folder structure, and also updated the
tests to use the fixtures folder for temporary test related file