Skip option for slow test locally
There are couple of tests that runs super slow, that is mainly
depends on the host for those formulas. So, unless we did any
changes on those formulas we don't necessarily breaking those

This commit adds a mechanism by adding a rspec tag `slow_test`,
so by default `bin/rspec` will ignore those slow tests. But it
will running the same as before in the CI environment.