Fix system font finder related issues
The system font finder was returning the whole list of fonts for
some cases like `Calib`, turns out it was an validation check for
the font names, it returns empty array & that's what was causing
this issue.

This commit fixes this issue, and also add a small check for this
so now we only return the matched list of fonts.