Add a time limit for the slow formulas
There are some formula that might be super slow sometime, and that
leaves the user with a feeling of stalling, so this commit adds a
time caps to these slow download. This commit set the connection
time out to 10s and read timeout to 10 seconds for each block of
response.

Fixes - 96
Link: https://github.com/ankane/the-ultimate-guide-to-ruby-timeouts