Add support for license agreement handling
Earlier, Fontist wasn't handling the proprietary license itself,
but relying on the dependent client library to handle it for us.
But we've decided to handle the required license by fontist, so
this commit adds the support for that.

There are couple of scenario: first of all, the open fonts, and
in those case fontist won't prompt the license agreement and it
will install the font from the correct formula.

Second scenario, proprietary font with required license, and in
this case fontist will display the license and ask the user to
provide their confirmation, and if it is same as the required
one then it will proceed or raise an error.

Fair warning: this might cause some issue if someone tries to
use fontist for some automation, where they don't provide the
correct license, but that's what we are trying to avoid and
want the client to handle it properly.

Fixes - 98